refactor(cli): decompose oversized lib modules, remove dead record-writers
sdd-utils.js (1267 lines) and coordinator.js (1070 lines) exceeded the
700-line hard limit. Both are now thin facades over cohesive submodules,
with export surfaces byte-identical to before — no call-site changes:

- sdd-utils.js -> sdd-spec-header.js, sdd-design-doc.js,
  sdd-decision-ledger.js (+ existing sdd-rules/sdd-validators re-exports)
- coordinator.js -> coordinator-core.js, coordinator-worktree-ops.js,
  coordinator-multi-spec.js. Worktree ops attach to the prototype from
  core; the three internal `new Coordinator(...)` calls become
  `new this.constructor(...)` so the dependency graph stays acyclic.

Delete recall-record-writer.js and reflect-record-writer.js: superseded
by operation-record-writer.js, which the recall/reflect flows and even
these files' own tests already use. Tests kept (they are the only
coverage for the live module); headers fixed to name the right module.
